chore: move howto categories under Howto section
This component was not a generic or reusable component instead it's
implementation is tightly coupled to the Howto entity

Closes https://github.com/ONEARMY/community-platform/issues/1716
